% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_function.R
\name{barchart_function}
\alias{barchart_function}
\title{Barchart of the Economic and Pollution variable for the European countries in a certain year}
\usage{
barchart_function(dataset = data, pol_var = "CO2 Emissions",
  eco_var = "Productivity in Primary Sector", yrs = 2000)
}
\arguments{
\item{pol_var}{A \code{list} of Pollution variables to select}

\item{eco_var}{A \code{list} of Economic variables to select}

\item{yrs}{A \code{list} of all the years to be chosen}
}
\value{
A \code{plot} containing a barchart
}
\description{
Produces a barchart with the chosen Economic and Pollution variable in order to view the standing of the
European countries in a chosen year.
}
\author{
Redwan Hasan
}
