% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_function.R
\name{bubble_function}
\alias{bubble_function}
\title{Bubble graph}
\usage{
bubble_function(x_var, y_var)
}
\arguments{
\item{x_var}{A \code{list} of variables to select}

\item{y_var}{A \code{list} of variables to select}
}
\value{
A \code{plot} representing two variables in a bubble chart
}
\description{
Data visualisation of two variables of the users choice
in a two dimensional plot.
A third dimension is represented by the size of the bubble
}
\examples{
bubble_function ('gdp','productivity')
}
\author{
Ana-Maria Casian
}
