% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_function.R
\name{comparison_function}
\alias{comparison_function}
\title{Interactive map of the European region showing the change between
Economic and Pollution variables over time}
\usage{
comparison_function(
  dataset = data,
  yrs = 2000,
  pol_var = "CO2 Emissions from Transport",
  eco_var = "GDP"
)
}
\arguments{
\item{yrs}{A \code{list} of years to select}

\item{pol_var}{A \code{list} of Pollution variables to select}

\item{eco_var}{A \code{list} of Economic variables to select}
}
\value{
\code{Plots} containing interactive maps of Europe based
on the variables chosen
}
\description{
Produces two interactive maps of the European region over
two different years to show the interaction
and change between the Economic and Pollution variables
}
\examples{
comparison_function(yrs = 2000,
pol_var = "CO2 Emissions from Transport", eco_var = "GDP")
}
\author{
Team Alpha
}
