% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_line_function.R
\name{country_function}
\alias{country_function}
\title{Variables per Country through Year}
\usage{
country_function(dataset = data, ctry = "Switzerland", var = "GDP")
}
\arguments{
\item{ctry}{A \code{list} of countries to select}

\item{var}{A \code{list} of variables to select}
}
\value{
A \code{plot} containing the line per countries for the selected
countries
}
\description{
Compute a linetype graph containing the desired variables per
desired countries through year. Multiple choices is allowed for the country
variable
}
\examples{
country_function(c('Italy', 'France', 'Spain'),'Productivity')
}
\author{
Team Alpha
}
