% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_function.R
\name{forecast_function}
\alias{forecast_function}
\title{Forecast of a variable over a number of years}
\usage{
forecast_function(dataset = data, geo = "France", variable = "GDP",
  fc_length = 5)
}
\arguments{
\item{dataset}{a preprocessed dataset from which information is extracted}

\item{geo}{a country that is contained in the dataset}

\item{variable}{a variable that is contained in the dataset}

\item{fc_length}{a integer representing the number of years of the forecast}
}
\value{
A \code{list} containing the time series created, the time series of
the forecast and the dygraph \code{plot}
}
\description{
Given data, a country, a variable and a forecasting length,
produces a time series and a forecast of the variable over time, and an
interactive plot of the forecast
}
\examples{
#store the function results in a variable
my_fc <- forecast_function(data, 'France', 'GDP', 10)
#plot the interactive dygraph plot
my_fc$plot


}
\author{
Pierre-Emmanuel Got
}
