% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generating the data for the ptdsAlpha package}
\usage{
generate_data()
}
\value{
'data'
}
\description{
Function that load the data
}
\examples{
generate_data()
}
\author{
Team Alpha
}
