% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif_function.R
\name{gif_function}
\alias{gif_function}
\title{Evolution of European indicators through time}
\usage{
gif_function(dataset = data, var = "CO2 Emissions")
}
\arguments{
\item{var}{A \code{list} of any variable to select from the dataset}
}
\value{
An \code{animation} containing the map of Europe based on the
variable chosen
}
\description{
Displays a GIF showing the evolution of one variable through
time over an EU chloropleth map
}
\examples{
gif_function('Greenhouse_Gas_Emissions')
}
\author{
Team Alpha
}
