% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_function.R
\name{map_function}
\alias{map_function}
\title{Pollution & Economic Overview of European countries}
\usage{
map_function(yrs, col_val)
}
\arguments{
\item{yrs}{A \code{list} of years to select}

\item{col_val}{A \code{list} of Economic and Pollution variables to select}
}
\value{
A \code{plot} containing the map of Europe based on the variable chosen
}
\description{
Produces a colorful graphical output of the European map in regards to the chosen Economic or Pollution variables
in a given year
}
\examples{
map_function(2008,'Productivity')
map_function(1982, 'Greenhouse_Gas_Emissions')
}
\author{
Redwan Hasan
}
