% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranktable_function.R
\name{ranktable_function}
\alias{ranktable_function}
\title{RankTable by Years, Countries and Variables}
\usage{
ranktable_function(
  dataset = data,
  yrs = "2000",
  ctry = "Switzerland",
  var = "GDP"
)
}
\arguments{
\item{yrs}{A \code{list} of Years to select}

\item{ctry}{A \code{list} of Countries to select}

\item{var}{A \code{list} of Variables to select}
}
\value{
A \code{table} containing the desired information and the rank for
each countries
}
\description{
Compute a table with the value of the selected variables for the
selected countries and years, and the rank
}
\examples{
ranktable_function(c(2001, 1992), c('Albania', 'Andorra'), c('GDP', 'Productivity'))
}
\author{
Claudio Previte
}
