% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_top10.R
\name{animated_top10}
\alias{animated_top10}
\title{Animated Top10 Countries per Variables}
\usage{
animated_top10(dataset = data, var = "Productivity")
}
\arguments{
\item{var}{A variable to select}
}
\value{
A \code{plot} containing an animated ranking
}
\description{
Compute an animated ranking with the 10 top European countries
per year in regards to a chosen economic or pollution variable
}
\examples{
animated_top10('Productivity')
}
\author{
Team Alpha
}
