% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_function.R
\name{bubble_function}
\alias{bubble_function}
\title{Bubble graph}
\usage{
bubble_function(
  dataset = data,
  ctry = "Switzerland",
  eco_var = "GDP",
  pol_var = "CO2 Emissions"
)
}
\arguments{
\item{x_var}{First variables to select}

\item{y_var}{Second variables to select}
}
\value{
A \code{plot} representing two variables in a bubble chart
}
\description{
Data visualisation of two variables of the users choice
in a two dimensional plot.
A third dimension is represented by the size of the bubble
}
\examples{
bubble_function ('gdp','productivity')
}
\author{
Ana-Maria Casian
}
