% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_function.R
\name{bubble_function}
\alias{bubble_function}
\title{Variables per Country through Year}
\usage{
bubble_function(x_var, y_var)
}
\arguments{
\item{ctry}{A \code{list} of countries to selectcoun}

\item{var}{A \code{list} of variables to select}
}
\value{
A \code{plot} containing the line per variables and per countries
}
\description{
Compute an linetype graph containing the desired variables per desired countries through year
}
\examples{
country_function(c('Italy', 'France', 'Spain'),'Productivity')
country_function('Italy', c('Productivity', 'GDP'))
}
\author{
Claudio Previte
}
