% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_function.R
\name{bubble_function}
\alias{bubble_function}
\title{Bubble graph}
\usage{
bubble_function(dataset = data, ctry = "Switzerland",
  eco_var = "GDP", pol_var = "CO2 Emissions")
}
\arguments{
\item{ctry}{a \code{list} of countries to select}

\item{eco_var}{Economic variable to select}

\item{pol_var}{Pollution variable to select}
}
\value{
A \code{plot} representing two variables in a bubble chart
}
\description{
Data visualisation of one economic variable agianst one pollution variable of the users choice
in a two dimensional plot, for the selected countries.
}
\examples{
bubble_function ('Switzerland','GDP','Productivity')
}
\author{
Team Alpha
}
