% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_function.R
\name{comparison_function}
\alias{comparison_function}
\title{Interactive map of the European region showing the change betweem Economic and Pollution variables over time}
\usage{
comparison_function(
  dataset = data,
  pol_var = "CO2 Emissions from Transport",
  eco_var = "GDP"
)
}
\arguments{
\item{pol_var}{A \code{list} of Pollution variables to select}

\item{eco_var}{A \code{list} of Economic variables to select}
}
\value{
A \code{plot} containing four interactive maps of Europe based on the variables chosen
}
\description{
Produces an interactive map of the European Union region over four different years to show the interaction
and change between the Economic and Pollution variables
}
\author{
Redwan Hasan
}
