% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif_function.R
\name{gif_function}
\alias{gif_function}
\title{Animation of the European indicators over time}
\usage{
gif_function(var)
}
\arguments{
\item{pol_var}{A \code{list} of any variable to select}
}
\value{
An \code{animation} containing the map of Europe based on the variable chosen
}
\description{
Produces a GIF about one variable and shows an EU chloropleth map with colors changing accordingly to the data evolution
}
\examples{
comparison_function('Greenhouse_Gas_Emissions')
comparison_function('CO2_Emissions_from_Transport')
}
\author{
Jeremy Choppe
}
