% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranktable_function.R
\name{ranktable_function}
\alias{ranktable_function}
\title{RankTable by Years, Countries and Variables}
\usage{
ranktable_function(dataset = data, yrs = c("1990", "2000"),
  ctry = c("France", "Switzerland"), var = c("CO2 Emissions", "GDP"))
}
\arguments{
\item{yrs}{A \code{list} of Years to select}

\item{ctry}{A \code{list} of Countries to select}

\item{var}{A \code{list} of Variables to select}
}
\value{
A \code{table} containing the desired information and the rank for
each countries
}
\description{
Compute a table with the value of the selected variables for the
selected countries and years, and the rank
}
\examples{
ranktable_function(
         yrs = c('2001', '1992'),
         ctry = c('Albania', 'Andorra'),
         var  = c('GDP', 'Productivity'))
}
\author{
Claudio Previte
}
